Refactor: Message -> InputView 와 ResultView 로 분리
